refactoring Models and Helpers Enums
