+++ Pokemon Challenge/Network/Enums/NetworkErrors.swift
@@ -1,0 +1,12 @@
+//
+//  NetworkErrors.swift
+//  Pokemon Challenge
+//
+//  Created by Evens Taian on 19/01/25.
+//
+
+enum NetworkErrors: Error {
+    case notFound
+    case noConnection
+    case unknown
+}
